requirements.txt
wheel
wget
minecraft_launcher_lib==4.0
pyfiglet
ttkbootstrap
tkvideo
requests
Pillow
tkhtmlview
speedtest-cli
